+++ argopy/data_fetchers/argovis_data.py
@@ -370,9 +370,8 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
         ds = self.filter_domain(ds)  # https://github.com/euroargodev/argopy/issues/48
         return ds
 
-    def filter_data_mode(self, ds: xr.Dataset, **kwargs):
-        # Argovis data already curated !
-        # ds = ds.argo.filter_data_mode(errors='ignore', **kwargs)
+    def transform_data_mode(self, ds: xr.Dataset, **kwargs):
+        # Argovis data are already curated !
         if ds.argo._type == "point":
             ds["N_POINTS"] = np.arange(0, len(ds["N_POINTS"]))
         return ds
@@ -389,7 +388,7 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
 
         This filter will select only QC=1 delayed mode data with pressure errors smaller than 20db
 
-        Use this filter instead of filter_data_mode and filter_qc
+        Use this filter instead of transform_data_mode and filter_qc
         """
         ds = ds.argo.filter_researchmode()
         if ds.argo._type == "point":