+++ argopy/data_fetchers/gdacftp_data.py
@@ -400,8 +400,8 @@ class FTPArgoDataFetcher(ArgoDataFetcherProto):
 
         return ds
 
-    def filter_data_mode(self, ds: xr.Dataset, **kwargs):
-        ds = ds.argo.filter_data_mode(errors="ignore", **kwargs)
+    def transform_data_mode(self, ds: xr.Dataset, **kwargs):
+        ds = ds.argo.transform_data_mode(params='all', errors="ignore", **kwargs)
         if ds.argo._type == "point":
             ds["N_POINTS"] = np.arange(0, len(ds["N_POINTS"]))
         return ds
@@ -417,7 +417,7 @@ class FTPArgoDataFetcher(ArgoDataFetcherProto):
 
             This filter will select only QC=1 delayed mode data with pressure errors smaller than 20db
 
-            Use this filter instead of filter_data_mode and filter_qc
+            Use this filter instead of transform_data_mode and filter_qc
         """
         ds = ds.argo.filter_researchmode()
         if ds.argo._type == "point":