+++ argopy/tests/test_fetchers_proto.py
@@ -10,8 +10,8 @@ class Fetcher(ArgoDataFetcherProto):
     def to_xarray(self, *args, **kwargs):
         super(Fetcher, self).to_xarray(*args, **kwargs)
 
-    def filter_data_mode(self, *args, **kwargs):
-        super(Fetcher, self).filter_data_mode(*args, **kwargs)
+    def transform_data_mode(self, *args, **kwargs):
+        super(Fetcher, self).transform_data_mode(*args, **kwargs)
 
     def filter_qc(self, *args, **kwargs):
         super(Fetcher, self).filter_qc(*args, **kwargs)
@@ -26,7 +26,7 @@ def test_required_methods():
         f.to_xarray()
 
     with pytest.raises(NotImplementedError):
-        f.filter_data_mode(xarray.Dataset, str)
+        f.transform_data_mode(xarray.Dataset, str)
 
     with pytest.raises(NotImplementedError):
         f.filter_qc(xarray.Dataset)